requirements.txt
Pillow==10.1.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
